requirements.txt
anthropic>=0.39.0
google-genai>=1.56.0
openai>=1.60.0
pypdf>=4.3.1
python-dotenv>=1.0.1
tqdm>=4.66.0
